+++ packages/framework/src/shared/actions.ts
@@ -60,6 +60,7 @@ export type NavigateModalScreenAction =
   | string
   | (NavigateScreenOptions & {
       maskClosable?: boolean;
+      mask?: boolean;
       hideFullScreenIcon?: boolean;
       hideCloseIcon?: boolean;
       title?: string | { [key: string]: unknown };
@@ -75,6 +76,7 @@ export type NavigateBackAction = null;
 export type NavigateExternalScreen = string | NavigateExternalScreenOptions;
 
 export interface ShowDialogOptions {
+  mask?: boolean;
   minWidth?: number;
   maxWidth?: number;
   minHeight?: number;
@@ -85,7 +87,8 @@ export interface ShowDialogOptions {
 }
 
 export interface ShowDialogAction {
-  widget: { [key: string]: unknown };
+  widget?: Record<string, unknown>;
+  body?: Record<string, unknown>;
   options?: ShowDialogOptions;
   onDialogDismiss?: EnsembleAction;
 }
@@ -118,11 +121,9 @@ export interface UploadFilesAction {
 
 export type CloseAllDialogsAction = null;
 
-/* eslint-disable tsdoc/syntax */
 /**
  * @uiType action
  */
-/* eslint-enable tsdoc/syntax */
 export type EnsembleAction =
   | {
       executeCode?: ExecuteCodeAction;