+++ packages/runtime/src/shared/coreSchema.ts
@@ -21,6 +21,9 @@ import type {
   TabBarProps,
   GridProps,
   TabBarItem,
+  CardProps,
+  FittedColumnProps,
+  FittedRowProps,
 } from "../widgets";
 import type { Menu, View } from "./screenSchema";
 import type { FlexboxProps, IconProps } from "./types";
@@ -189,6 +192,7 @@ export type Widget =
   | { Button?: ButtonProps }
   // | { IconButton?: IconButton }
   // | { Address?: Address }
+  | { Card?: ReplaceChildrenTemplate<CardProps> }
   | {
       Column?: ReplaceChildrenTemplate<FlexboxProps>;
     }
@@ -208,8 +212,8 @@ export type Widget =
   // | { Calendar?: Calendar }
   // | { Countdown?: Countdown }
   | { ToggleButton?: ToggleButtonProps }
-  // | { FittedRow?: FittedRow }
-  // | { FittedColumn?: FittedColumn }
+  | { FittedRow?: ReplaceChildrenTemplate<FittedRowProps> }
+  | { FittedColumn?: ReplaceChildrenTemplate<FittedColumnProps> }
   // | { Map?: Map }
   | { Carousel?: ReplaceChildrenTemplate<CarouselProps> }
   // | { Conditional?: ConditionalProps };