+++ packages/runtime/src/widgets/Button.tsx
@@ -12,10 +12,12 @@ export type ButtonProps = {
   onTap?: EnsembleAction;
   submitForm?: boolean;
   startingIcon?: IconProps;
+  disabled?: boolean;
   endingIcon?: IconProps;
   styles?: {
-    textColor: string;
-    gap: string | number;
+    /** @uiType color */
+    textColor?: string;
+    gap?: string | number;
   };
 } & EnsembleWidgetProps;
 
@@ -36,6 +38,7 @@ export const Button: React.FC<ButtonProps> = ({ id, onTap, ...rest }) => {
   const ButtonComponent = useMemo(() => {
     return (
       <AntButton
+        disabled={values?.disabled ?? false}
         htmlType="submit"
         onClick={onClickCallback}
         ref={rootRef}